ch: make register view its own view
